Add documentation to avoid absolute path problem
Normally git uses relative paths for internal references. However, the
stock version of git in OS X El Capitan, "2.7.4 (Apple Git-66)", will
use absolute paths for submodule's gitdir property when the submodule
is initialized at the same time as the parent repository is.

The workaround is to clone the repository in two steps, so that the
submodule is initialized after the parent repository and given relative
paths like Vagrant expects.